# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/eeg/demandes":{"post":{"operationId":"DemandesController_creerDemande","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"patientId":{"type":"string","example":"4b0f2a29-af9b-4e93-9db6-78c27d19c30c"},"episodeSoinsId":{"type":"string","example":"EPI-001"},"typeEEG":{"type":"string","enum":["STANDARD","SOMMEIL","AMBULATOIRE","VIDEO_EEG"],"example":"STANDARD"},"urgence":{"type":"string","enum":["STAT","URGENTE","NORMALE"],"example":"URGENTE"},"motifPrescription":{"type":"string","example":"Motif de la prescription"}},"required":["patientId","episodeSoinsId","typeEEG","urgence","motifPrescription"]}}}},"responses":{"201":{"description":""}},"summary":"Créer une nouvelle demande EEG","tags":["Demandes"]}},"/eeg/demandes/worklist":{"get":{"operationId":"DemandesController_getWorklist","parameters":[{"name":"role","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Worklist filtrée par rôle","tags":["Demandes"]}},"/eeg/demandes/patient/{patientId}":{"get":{"operationId":"DemandesController_getDemandesByPatient","parameters":[{"name":"patientId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Historique EEG d'un patient","tags":["Demandes"]}},"/eeg/demandes/{id}":{"get":{"operationId":"DemandesController_getDemandeById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Détail d'une demande EEG","tags":["Demandes"]}},"/eeg/demandes/{id}/valider":{"patch":{"operationId":"DemandesController_validerDemande","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"MEDECIN_SERVICE : Valider une demande (CREEE → VALIDEE)","tags":["Demandes"]}},"/eeg/demandes/{id}/annuler":{"patch":{"operationId":"DemandesController_annulerDemande","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Annuler une demande EEG","tags":["Demandes"]}},"/eeg/demandes/{id}/refuser":{"patch":{"operationId":"DemandesController_refuserDemande","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"MEDECIN_SERVICE : Refuser une demande avec motif","tags":["Demandes"]}},"/eeg/demandes/{id}/planifier":{"patch":{"operationId":"DemandesController_planifierRdv","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"CHEF_SERVICE : Planifier un RDV (VALIDEE → PLANIFIEE)","tags":["Demandes"]}},"/eeg/demandes/{id}/realiser":{"patch":{"operationId":"DemandesController_realiserDemande","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"TECHNICIEN : Réaliser un examen","tags":["Demandes"]}},"/eeg/demandes/{id}/interpreter":{"patch":{"operationId":"DemandesController_interpreterDemande","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"MEDECIN_SERVICE : Remplir le brouillon CR","tags":["Demandes"]}},"/eeg/demandes/{id}/valider-cr":{"patch":{"operationId":"DemandesController_validerCR","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"CHEF_SERVICE : Valider le compte rendu final","tags":["Demandes"]}},"/eeg/demandes/{id}/ack":{"patch":{"operationId":"DemandesController_accuserReception","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Accusé de réception du résultat","tags":["Demandes"]}},"/eeg/upload/image/{demandeId}":{"post":{"operationId":"ResultatsController_uploadImageTrace","parameters":[{"name":"demandeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Resultats"]}},"/eeg/rapports/activite":{"get":{"operationId":"RapportsController_getActivite","parameters":[],"responses":{"200":{"description":""}},"summary":"Volumes des demandes EEG","tags":["Rapports"]}},"/eeg/rapports/delais":{"get":{"operationId":"RapportsController_getDelais","parameters":[],"responses":{"200":{"description":""}},"summary":"Délais moyens de traitement","tags":["Rapports"]}},"/eeg/rapports/anomalies":{"get":{"operationId":"RapportsController_getAnomalies","parameters":[],"responses":{"200":{"description":""}},"summary":"Résultats critiques non ACK","tags":["Rapports"]}},"/eeg/archives":{"get":{"description":"Retourne uniquement les résultats validés (estImmutable=true). Filtrable par patient, date, type EEG, conclusion, numéro EEG.","operationId":"ArchivesController_getArchives","parameters":[{"name":"patientId","required":false,"in":"query","schema":{"type":"string"}},{"name":"numeroEEG","required":false,"in":"query","schema":{"type":"string"}},{"name":"typeEEG","required":false,"in":"query","schema":{"enum":["STANDARD","SOMMEIL","AMBULATOIRE","VIDEO_EEG"],"type":"string"}},{"name":"dateDebut","required":false,"in":"query","description":"YYYY-MM-DD","schema":{"type":"string"}},{"name":"dateFin","required":false,"in":"query","description":"YYYY-MM-DD","schema":{"type":"string"}},{"name":"conclusion","required":false,"in":"query","description":"Recherche texte libre dans conclusionDiagnostique","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","description":"Page (défaut 1)","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Résultats par page (défaut 20, max 100)","schema":{"type":"string"}}],"responses":{"200":{"description":"Liste paginée des résultats archivés"}},"summary":"Archives EEG — résultats validés","tags":["Archives"]}},"/eeg/audit":{"get":{"description":"Retourne les 100 dernières entrées d'audit, triées par date décroissante. Filtrable par patientId, demandeId ou action.","operationId":"AuditController_getAudit","parameters":[{"name":"patientId","required":false,"in":"query","description":"Filtrer par identifiant patient","schema":{"type":"string"}},{"name":"demandeId","required":false,"in":"query","description":"Filtrer par identifiant demande","schema":{"type":"string"}},{"name":"action","required":false,"in":"query","description":"Filtrer par type d'action","schema":{"enum":["CREATION","CONSULTATION","MODIFICATION","VALIDATION","ANNULATION","IMPRESSION","ACK"],"type":"string"}}],"responses":{"200":{"description":"Journal d'audit (max 100 entrées)"}},"summary":"Journal d'audit EEG","tags":["Audit"]}},"/eeg/notifications":{"get":{"operationId":"NotificationsController_getNotifications","parameters":[{"name":"lu","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Notifications"]}},"/eeg/notifications/count":{"get":{"operationId":"NotificationsController_countNonLues","parameters":[],"responses":{"200":{"description":""}},"tags":["Notifications"]}},"/eeg/notifications/{id}":{"get":{"operationId":"NotificationsController_getNotificationById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Notifications"]}},"/eeg/notifications/{id}/lu":{"patch":{"operationId":"NotificationsController_marquerCommeLue","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Notifications"]}},"/eeg/notifications/lire-tout":{"patch":{"operationId":"NotificationsController_marquerToutesCommeLues","parameters":[],"responses":{"200":{"description":""}},"tags":["Notifications"]}},"/eeg/rdvs":{"get":{"operationId":"RdvsController_getRdvs","parameters":[{"name":"statut","required":true,"in":"query","schema":{"type":"string"}},{"name":"dateDebut","required":true,"in":"query","schema":{"type":"string"}},{"name":"dateFin","required":true,"in":"query","schema":{"type":"string"}},{"name":"patientId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Rdvs"]},"post":{"operationId":"RdvsController_creerRdv","parameters":[],"responses":{"201":{"description":""}},"tags":["Rdvs"]}},"/eeg/rdvs/semaine":{"get":{"operationId":"RdvsController_getRdvsSemaine","parameters":[],"responses":{"200":{"description":""}},"tags":["Rdvs"]}},"/eeg/rdvs/today":{"get":{"operationId":"RdvsController_getRdvsAujourdhui","parameters":[],"responses":{"200":{"description":""}},"tags":["Rdvs"]}},"/eeg/rdvs/{id}":{"get":{"operationId":"RdvsController_getRdvById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Rdvs"]},"patch":{"operationId":"RdvsController_modifierRdv","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Rdvs"]},"delete":{"operationId":"RdvsController_supprimerRdv","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Rdvs"]}},"/eeg/rdvs/{id}/realiser":{"patch":{"operationId":"RdvsController_realiserRdv","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Rdvs"]}},"/eeg/rdvs/{id}/non-realise":{"patch":{"operationId":"RdvsController_marquerNonRealise","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Rdvs"]}},"/eeg/rdvs/{id}/annuler":{"patch":{"operationId":"RdvsController_annulerRdv","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Rdvs"]}},"/eeg/patients":{"get":{"operationId":"PatientsController_getPatients","parameters":[{"name":"search","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Patients"]},"post":{"operationId":"PatientsController_creerPatient","parameters":[],"responses":{"201":{"description":""}},"tags":["Patients"]}},"/eeg/patients/dossier/{idDossier}":{"get":{"operationId":"PatientsController_getPatientByDossier","parameters":[{"name":"idDossier","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Patients"]}},"/eeg/patients/{id}":{"get":{"operationId":"PatientsController_getPatientById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Patients"]},"patch":{"operationId":"PatientsController_modifierPatient","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Patients"]}}},"info":{"title":"SIH CHUA — Module EEG","description":"API du module Électroencéphalographie — CHU Andrainjato Fianarantsoa","version":"1.0","contact":{}},"tags":[{"name":"Demandes","description":"Gestion des demandes EEG"},{"name":"Résultats","description":"Upload, validation et rectification des résultats"},{"name":"Rapports","description":"Rapports et statistiques"},{"name":"Audit","description":"Journal des actions"}],"servers":[],"components":{"schemas":{}}}